backend/requirements.txt
# ============================================================================
# Repazoo Backend Requirements
# FastAPI + Supabase + OAuth + Stripe + Redis Stack
# ============================================================================

# ============================================================================
# Core Framework
# ============================================================================
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# ============================================================================
# Database (Supabase/PostgreSQL)
# ============================================================================
supabase==2.3.4

# ============================================================================
# Authentication & Security
# ============================================================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
cryptography==42.0.2

# ============================================================================
# Payment Processing (Stripe)
# ============================================================================
stripe==7.9.0

# ============================================================================
# Redis (Rate Limiting & Caching)
# ============================================================================
redis==5.0.1

# ============================================================================
# Environment & Configuration
# ============================================================================
python-dotenv==1.0.1

# ============================================================================
# Data Serialization
# ============================================================================
orjson==3.9.15

# ============================================================================
# Date/Time Utilities
# ============================================================================
python-dateutil==2.8.2

# ============================================================================
# Monitoring & Logging (Optional)
# ============================================================================
# Sentry SDK for error tracking (uncomment if using Sentry)
# sentry-sdk==1.39.2

# ============================================================================
# API Documentation
# ============================================================================
starlette==0.36.3

# ============================================================================
# Production Server
# ============================================================================
gunicorn==21.2.0

# ============================================================================
# Testing Dependencies (Development)
# ============================================================================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# ============================================================================
# Code Quality (Development)
# ============================================================================
black==23.12.1
ruff==0.1.9
mypy==1.8.0

# ============================================================================
# Additional Utilities
# ============================================================================
# Task queue (if needed for async analysis)
# celery==5.3.4
# celery[redis]==5.3.4

# AI/ML Libraries (if running analysis in backend)
# anthropic==0.8.1  # Anthropic Claude API client
